api/requirements.txt
fastapi
pydantic
python-dotenv
openai
google-generativeai
tiktoken
requests
httpx
